+++ feature/src/main/AndroidManifest.xml
