+++ settings.gradle.kts
@@ -22,4 +22,5 @@ dependencyResolutionManagement {
 
 rootProject.name = "Lothar News"
 include(":app")
+include(":feature")
 include(":ui")
